pyproject.toml
[tool.poetry]
name = "dash-ecomm"
version = "0.1.0"
description = "An e-comm bot for customer demos"
authors = ["Ayushman Dash <ayushman@neuralspace.ai>", "Dishant Gandhi <dishant@neuralspace.ai>"]


[[tool.poetry.source]]
name = "rasax"
url = "https://pypi.rasa.com/simple"

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
rasa = "^2.3.4"
rasa-x = "^0.37.1"
PyYAML = "^5.4.1"
python-socketio = "4.6.1"
python-engineio = "3.13.2"

[tool.poetry.dev-dependencies]
isort = "^4.3.21"
flake8 = "^3.7.9"
pytest = "^5.4.1"
coverage = "^5.1"
requirements-parser = "^0.2.0"
pytest-xdist = "^1.32.0"
black = "^20.8b1"
twine = "^3.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
